+++ src/GRA.Controllers/ViewModel/MissionControl/Schools/EnteredListViewModel.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.ComponentModel;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Schools
 {
@@ -11,7 +12,10 @@ namespace GRA.Controllers.ViewModel.MissionControl.Schools
     {
         public List<GRA.Domain.Model.EnteredSchool> EnteredSchools { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
+        public bool NewSchool { get; set; }
         public GRA.Domain.Model.School EnteredSchool { get; set; }
+        [DisplayName("School")]
+        public int SchoolId { get; set; }
         public SelectList SchoolDistricts { get; set; }
         public SelectList SchoolTypes { get; set; }
     }